+++ src/main.jsx
@@ -5,25 +5,20 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import "./assets/css/index.css";
 
 // Layouts
-import App from "./App.jsx";
-import DoctorApp from "./pages/DoctorApp/DoctorApp.jsx";
+import App from "./pages/Layout/SecratariaGApp.jsx";           // Layout Admin
+import DoctorApp from "./pages/Layout/DoctorApp.jsx"; // Layout Médico
 
-// Página de Login
-import Login from "./pages/Login/Login.jsx";
-
-import SecretariaDashboard from "./pages/Secretaria/SecretariaDashboard.jsx";
 // Páginas Admin
 import Patientform from "./pages/Patient/Patientform.jsx";
 import PatientList from "./pages/Patient/PatientList.jsx";
-import Doctorlist from "./pages/Doctor/DoctorList.jsx";
-import DoctorForm from "./pages/Doctor/DoctorForm.jsx";
+import Doctorlist from "./pages/Secretaria Geral/Doctor/DoctorList.jsx";
+import DoctorForm from "./pages/Secretaria Geral/Doctor/DoctorForm.jsx";
 import Doctorschedule from "./pages/Schedule/DoctorSchedule.jsx";
 import AddSchedule from "./pages/Schedule/AddSchedule.jsx";
-import Calendar from "./pages/calendar/Calendar.jsx";
-import EditDoctor from "./pages/Doctor/DoctorEdit.jsx";
+import Calendar from "./pages/Secretaria Geral/calendar/Calendar.jsx";
+import EditDoctor from "./pages/Secretaria Geral/Doctor/DoctorEdit.jsx";
 import PatientEdit from "./pages/Patient/PatientEdit.jsx";
-import DoctorProfile from "./pages/Doctor/DoctorProfile.jsx";
-
+import DoctorProfile from "./pages/Secretaria Geral/Doctor/DoctorProfile.jsx";
 
 
 // Páginas Médico
@@ -31,36 +26,24 @@ import DoctorDashboard from "./pages/DoctorApp/DoctorDashboard.jsx";
 import DoctorCalendar from "./pages/DoctorApp/DoctorCalendar.jsx";
 import DoctorPatientList from "./pages/DoctorApp/Patient/DoctorPatientList.jsx";
 import ConsultaList from "./pages/DoctorApp/Consultas/ConsultaList.jsx";
-
-
-import AgendaList from "./pages/Agendar/AgendaList.jsx";
-import AgendaForm from "./pages/Agendar/AgendaForm.jsx";
-import AgendaEdit from "./pages/Agendar/AgendaEdit.jsx";
-import LaudoList from "./pages/laudos/LaudosList.jsx";
+import AgendaList from "./pages/Secretaria Geral/Agendar/AgendaList.jsx";
+import AgendaForm from "./pages/Secretaria Geral/Agendar/AgendaForm.jsx";
+import AgendaEdit from "./pages/Secretaria Geral/Agendar/AgendaEdit.jsx";
+import LaudoList from "./pages/laudos/LaudosList.jsx"
 import Laudo from "./pages/laudos/Laudo.jsx";
 
-// Páginas Médico
-import DoctorDashboard from "./pages/DoctorApp/DoctorDashboard.jsx";
-import DoctorCalendar from "./pages/DoctorApp/DoctorCalendar.jsx";
-import DoctorPatientList from "./pages/DoctorApp/DoctorPatientList.jsx";
 
-// Router
+
+
+// Criando o router com todas as rotas
 const router = createBrowserRouter([
-  // Primeira tela = Login
-  {
-    path: "/",
-    element: <Login />,
-  },
-
-  {
-    path: "/secretaria",
-    element: <SecretariaDashboard />,
-  },
-  
+   
+  // Rotas Admin
   {
     path: "/",
     element: <App />,
     children: [
+      // Rota inicial do Admin: apenas mostra layout com Navbar e Sidebar
       { path: "patient", element: <Patientform /> },
       { path: "patientlist", element: <PatientList /> },
       { path: "doctorlist", element: <Doctorlist /> },
@@ -71,20 +54,19 @@ const router = createBrowserRouter([
       { path: "profiledoctor/:id", element: <DoctorProfile /> },
       { path: "editdoctor/:id", element: <EditDoctor /> },
       { path: "editpatient/:id", element: <PatientEdit /> },
-      { path: "agendaform", element: <AgendaForm /> },
-      { path: "agendaedit", element: <AgendaEdit /> },
-      { path: "agendalist", element: <AgendaList /> },
+      { path: "agendaform", element: <AgendaForm />},
+      { path: "agendaedit", element: <AgendaEdit />},
+      { path: "agendalist", element: <AgendaList />},
       { path: "laudolist", element: <LaudoList /> },
-      { path: "laudo", element: <Laudo /> },
+      { path: "laudo", element: <Laudo />}
     ],
   },
-
   // Rotas Médico
   {
     path: "/doctor",
     element: <DoctorApp />,
     children: [
-      { index: true, element: <DoctorDashboard /> },
+      { index: true, element: <DoctorDashboard /> }, // Rota inicial médico
       { path: "dashboard", element: <DoctorDashboard /> },
       { path: "calendar", element: <DoctorCalendar /> },
       { path: "patients", element: <DoctorPatientList /> },
@@ -93,6 +75,7 @@ const router = createBrowserRouter([
   },
 ]);
 
+// Renderizando a aplicação
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <RouterProvider router={router} />
