+++ src/pages/Layout/SecratariaGApp.jsx
@@ -1,0 +1,18 @@
+// src/App.jsx
+import './../../assets/css/index.css'
+import Navbar from './../../components/Navbar'
+import Sidebar from './../../components/Sidebar'
+import { Outlet } from 'react-router-dom'
+import SecretariaDashboard from './../../pages/Secretaria/SecretariaDashboard'
+
+function App() {
+  return (
+    <div>
+      <Navbar />
+      <Sidebar />
+      <Outlet />
+    </div>
+  )
+}
+
+export default App
