+++ src/pages/Secretaria Geral/Agendar/AgendaList.jsx
@@ -1,4 +1,4 @@
-import "../../assets/css/index.css"
+import "../../../assets/css/index.css";
 import { Link } from "react-router-dom";
 import { useState, useEffect, useRef, useLayoutEffect } from "react";
 import { createPortal } from "react-dom";
